refactor: extract SettingsResolver from Program.cs
